Share refined common tests
